# swagger-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["Index"]}},"/api/v1/auth/login":{"post":{"operationId":"UserAuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequestDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerBaseApiResponseFor class LoginResponseDto {\n} "}}}}},"tags":["Auth"]}},"/api/v1/auth/user/sign-up":{"post":{"operationId":"UserAuthController_signup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpRequestDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerBaseApiResponseFor class UserResponseDto {\n} "}}}}},"tags":["Auth"]}},"/api/v1/auth/admin/sign-up":{"post":{"operationId":"UserAuthController_adminSignup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRequestDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerBaseApiResponseFor class UserResponseDto {\n} "}}}}},"tags":["Auth"]}},"/api/v1/groups":{"post":{"operationId":"GroupController_createGroup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupRequestDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerBaseApiResponseFor class GroupResponseDto {\n} "}}}}},"tags":["Group(Roles)"]},"get":{"operationId":"GroupController_findGroup","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerBaseApiResponseFor class GroupResponseDto {\n}  [ ] "}}}}},"tags":["Group(Roles)"]}},"/api/v1/permissions":{"post":{"operationId":"PermissionController_createPermission","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PermissionRequestDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerBaseApiResponseFor class PermissionResponseDto {\n} "}}}}},"tags":["Permission"],"security":[{"admin-auth":[]}]},"get":{"operationId":"PermissionController_findPermission","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerBaseApiResponseFor class PermissionResponseDto {\n}  [ ] "}}}}},"tags":["Permission"],"security":[{"admin-auth":[]}]}},"/api/v1/permissions/assign":{"post":{"operationId":"PermissionController_assignPermission","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MapPermissionRequestDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerBaseApiResponseFor function String() { [native code] } "}}}}},"tags":["Permission"],"security":[{"admin-auth":[]}]}},"/api/v1/permissions/un-assign":{"delete":{"operationId":"PermissionController_unAssignPermission","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MapPermissionRequestDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerBaseApiResponseFor function String() { [native code] } "}}}}},"tags":["Permission"],"security":[{"admin-auth":[]}]}},"/api/v1/quizzes":{"post":{"operationId":"QuizController_createQuiz","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizRequestDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerBaseApiResponseFor function String() { [native code] } "}}}}},"tags":["Quiz"],"security":[{"admin-auth":[]}]},"get":{"operationId":"QuizController_findQuiz","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerBaseApiResponseFor class QuizResponseDto {\n}  [ ] "}}}}},"tags":["Quiz"]}},"/api/v1/quizzes/submission":{"post":{"operationId":"QuizController_submission","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuizSubmissionRequestDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerBaseApiResponseFor function String() { [native code] } "}}}}},"tags":["Quiz"],"security":[{"auth":[]}]}},"/api/v1/quizzes/result-metrics":{"get":{"operationId":"QuizController_quizResultMetrics","parameters":[{"name":"range_start_at","required":false,"in":"query","schema":{"format":"date-time","type":"string"}},{"name":"range_end_at","required":false,"in":"query","schema":{"format":"date-time","type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerBaseApiResponseFor class QuizResultMetric {\n}  [ ] "}}}}},"tags":["Quiz"],"security":[{"admin-auth":[]}]}},"/api/v1/quizzes/result-metrics/event":{"get":{"operationId":"QuizController_sendQuizResultMetricsEvent","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwaggerBaseApiResponseFor class QuizResultMetric {\n}  [ ] "}}}}},"tags":["Quiz"],"security":[{"admin-auth":[]}]}},"/api/v1/quizzes/stop-sending-event":{"get":{"operationId":"QuizController_stopEvents","parameters":[],"responses":{"200":{"description":""}},"tags":["Quiz"],"security":[{"admin-auth":[]}]}}},"info":{"title":"Pet Project","description":"API DOC","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"auth":{"type":"apiKey","in":"header","name":"Authorization","scheme":"bearer","bearerFormat":"Bearer"},"admin-auth":{"type":"apiKey","in":"header","name":"Authorization","scheme":"bearer","bearerFormat":"Bearer"}},"schemas":{"LoginRequestDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"LoginResponseDto":{"type":"object","properties":{"token":{"type":"string"}},"required":["token"]},"SwaggerBaseApiResponseFor class LoginResponseDto {\n} ":{"type":"object","properties":{"status":{"type":"number","default":200},"message":{"type":"string"},"errors":{"type":"array","items":{"type":"string"}},"payload":{"type":"object"},"data":{"$ref":"#/components/schemas/LoginResponseDto"}},"required":["status","message","errors","payload","data"]},"SignUpRequestDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"}},"required":["name","email","password"]},"UserResponseDto":{"type":"object","properties":{"_id":{"type":"string"},"name":{"type":"string"},"email":{"type":"string"}},"required":["_id","name","email"]},"SwaggerBaseApiResponseFor class UserResponseDto {\n} ":{"type":"object","properties":{"status":{"type":"number","default":200},"message":{"type":"string"},"errors":{"type":"array","items":{"type":"string"}},"payload":{"type":"object"},"data":{"$ref":"#/components/schemas/UserResponseDto"}},"required":["status","message","errors","payload","data"]},"UserRequestDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"group_id":{"type":"string"}},"required":["name","email","password","group_id"]},"GroupRequestDto":{"type":"object","properties":{"name":{"type":"string"},"identifier":{"type":"string","default":"user||admin"}},"required":["name","identifier"]},"GroupResponseDto":{"type":"object","properties":{"_id":{"type":"string"},"name":{"type":"string"},"identifier":{"type":"string"}},"required":["_id","name","identifier"]},"SwaggerBaseApiResponseFor class GroupResponseDto {\n} ":{"type":"object","properties":{"status":{"type":"number","default":201},"message":{"type":"string"},"errors":{"type":"array","items":{"type":"string"}},"payload":{"type":"object"},"data":{"$ref":"#/components/schemas/GroupResponseDto"}},"required":["status","message","errors","payload","data"]},"SwaggerBaseApiResponseFor class GroupResponseDto {\n}  [ ] ":{"type":"object","properties":{"status":{"type":"number","default":200},"message":{"type":"string"},"errors":{"type":"array","items":{"type":"string"}},"payload":{"type":"object"},"data":{"type":"array","items":{"$ref":"#/components/schemas/GroupResponseDto"}}},"required":["status","message","errors","payload","data"]},"PermissionRequestDto":{"type":"object","properties":{"name":{"type":"string","description":"Create Quiz"},"identifier":{"type":"string","default":"quiz.create||quiz.view||quiz.delete||quiz.result"},"status":{"type":"string","default":"active||in_active"}},"required":["name","identifier","status"]},"PermissionResponseDto":{"type":"object","properties":{"_id":{"type":"string"},"name":{"type":"string"},"identifier":{"type":"string"},"status":{"type":"string"}},"required":["_id","name","identifier","status"]},"SwaggerBaseApiResponseFor class PermissionResponseDto {\n} ":{"type":"object","properties":{"status":{"type":"number","default":201},"message":{"type":"string"},"errors":{"type":"array","items":{"type":"string"}},"payload":{"type":"object"},"data":{"$ref":"#/components/schemas/PermissionResponseDto"}},"required":["status","message","errors","payload","data"]},"SwaggerBaseApiResponseFor class PermissionResponseDto {\n}  [ ] ":{"type":"object","properties":{"status":{"type":"number","default":200},"message":{"type":"string"},"errors":{"type":"array","items":{"type":"string"}},"payload":{"type":"object"},"data":{"type":"array","items":{"$ref":"#/components/schemas/PermissionResponseDto"}}},"required":["status","message","errors","payload","data"]},"MapPermissionRequestDto":{"type":"object","properties":{"user_id":{"type":"string"},"permission_id":{"type":"string"}},"required":["user_id","permission_id"]},"SwaggerBaseApiResponseFor function String() { [native code] } ":{"type":"object","properties":{"status":{"type":"number","default":200},"message":{"type":"string"},"errors":{"type":"array","items":{"type":"string"}},"payload":{"type":"object"},"data":{"type":"string"}},"required":["status","message","errors","payload","data"]},"QuizOption":{"type":"object","properties":{"title":{"type":"string"},"is_correct":{"type":"boolean"}},"required":["title","is_correct"]},"QuizRequestDto":{"type":"object","properties":{"title":{"type":"string"},"status":{"type":"string","default":"active"},"options":{"type":"array","items":{"$ref":"#/components/schemas/QuizOption"}}},"required":["title","status","options"]},"QuizOptionResponse":{"type":"object","properties":{"title":{"type":"string"},"_id":{"type":"string"}},"required":["title","_id"]},"QuizResponseDto":{"type":"object","properties":{"_id":{"type":"string"},"title":{"type":"string"},"options":{"type":"array","items":{"$ref":"#/components/schemas/QuizOptionResponse"}}},"required":["_id","title","options"]},"SwaggerBaseApiResponseFor class QuizResponseDto {\n}  [ ] ":{"type":"object","properties":{"status":{"type":"number","default":200},"message":{"type":"string"},"errors":{"type":"array","items":{"type":"string"}},"payload":{"type":"object"},"data":{"type":"array","items":{"$ref":"#/components/schemas/QuizResponseDto"}}},"required":["status","message","errors","payload","data"]},"QuizSubmissionRequestDto":{"type":"object","properties":{"quiz_id":{"type":"string"},"option_id":{"type":"string"}},"required":["quiz_id","option_id"]},"QuizResultMetric":{"type":"object","properties":{"_id":{"type":"string"},"title":{"type":"string"},"total_submissions":{"type":"number"},"total_wrong_submissions":{"type":"number"},"total_correct_submissions":{"type":"number"},"percentage_of_wrong_submissions":{"type":"number"},"percentage_of_correct_submissions":{"type":"number"}},"required":["_id","title","total_submissions","total_wrong_submissions","total_correct_submissions","percentage_of_wrong_submissions","percentage_of_correct_submissions"]},"SwaggerBaseApiResponseFor class QuizResultMetric {\n}  [ ] ":{"type":"object","properties":{"status":{"type":"number","default":200},"message":{"type":"string"},"errors":{"type":"array","items":{"type":"string"}},"payload":{"type":"object"},"data":{"type":"array","items":{"$ref":"#/components/schemas/QuizResultMetric"}}},"required":["status","message","errors","payload","data"]}}}}